mts7/JAT-11 - consolidated to common layout, updated tests, changed routing
