+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"compilerOptions": {
-		"module": "es2015",
+		"module": "esnext",
 		"strict": true,
 		"target": "es5",
 		"sourceMap": true,
